Minor reorganisation
