+++ h5glance/__init__.py
@@ -1,4 +1,4 @@
 """Explore HDF5 files in an HTML view"""
-from .generate import H5Glance, install_ipython_h5py_display
+from .ipython import H5Glance, install_ipython_h5py_display
 
 __version__ = "0.7"
